Configure MySQL connection for course-service on GCP
This commit updates the course-service to use Google Cloud Platform (GCP) MySQL database settings. It sets the database host to the GCP instance IP, configures the port, JDBC URL, username, and password in the application-gcp.properties file. This enables the service to connect securely to the cloud-hosted MySQL database for course data management, supporting deployment and integration in the GCP environment.